Refactor MCP rules YAML helpers
